+++ sklearn/cluster/bicluster.py
@@ -15,7 +15,7 @@ from scipy.sparse.linalg import eigsh, svds
 from . import KMeans, MiniBatchKMeans
 from ..base import BaseEstimator, BiclusterMixin
 from ..externals import six
-from ..utils import _init_arpack_v0
+from ..utils.arpack import _init_arpack_v0
 
 from ..utils.extmath import (make_nonnegative, randomized_svd,
                              safe_sparse_dot)
