+++ sklearn/cross_decomposition/pls_.py
@@ -13,7 +13,8 @@ from scipy.linalg import pinv2, svd
 from scipy.sparse.linalg import svds
 
 from ..base import BaseEstimator, RegressorMixin, TransformerMixin
-from ..utils import check_array, check_consistent_length, _init_arpack_v0
+from ..utils import check_array, check_consistent_length
+from ..utils.arpack import _init_arpack_v0
 from ..utils.extmath import svd_flip
 from ..utils.validation import check_is_fitted, FLOAT_DTYPES
 from ..exceptions import ConvergenceWarning
