+++ sklearn/decomposition/kernel_pca.py
@@ -7,7 +7,7 @@ import numpy as np
 from scipy import linalg
 from scipy.sparse.linalg import eigsh
 
-from ..utils import _init_arpack_v0
+from ..utils.arpack import _init_arpack_v0
 from ..utils.validation import check_is_fitted, check_array
 from ..exceptions import NotFittedError
 from ..base import BaseEstimator, TransformerMixin
