+++ sklearn/decomposition/pca.py
@@ -24,7 +24,7 @@ from ..externals import six
 from .base import _BasePCA
 from ..utils import check_random_state
 from ..utils import check_array
-from ..utils import _init_arpack_v0
+from ..utils.arpack import _init_arpack_v0
 from ..utils.extmath import fast_logdet, randomized_svd, svd_flip
 from ..utils.extmath import stable_cumsum
 from ..utils.validation import check_is_fitted
