+++ sklearn/decomposition/truncated_svd.py
@@ -11,7 +11,8 @@ import scipy.sparse as sp
 from scipy.sparse.linalg import svds
 
 from ..base import BaseEstimator, TransformerMixin
-from ..utils import check_array, check_random_state, _init_arpack_v0
+from ..utils import check_array, check_random_state
+from ..utils.arpack import _init_arpack_v0
 from ..utils.extmath import randomized_svd, safe_sparse_dot, svd_flip
 from ..utils.sparsefuncs import mean_variance_axis
