+++ sklearn/manifold/locally_linear.py
@@ -10,7 +10,8 @@ from scipy.sparse import eye, csr_matrix
 from scipy.sparse.linalg import eigsh
 
 from ..base import BaseEstimator, TransformerMixin
-from ..utils import check_random_state, check_array, _init_arpack_v0
+from ..utils import check_random_state, check_array
+from ..utils.arpack import _init_arpack_v0
 from ..utils.extmath import stable_cumsum
 from ..utils.validation import check_is_fitted
 from ..utils.validation import FLOAT_DTYPES
