+++ sklearn/utils/tests/test_utils.py
@@ -275,15 +275,3 @@ def test_get_chunk_n_rows(row_bytes, max_n_rows, working_memory,
                                            ([np.nan], False)])
 def test_is_scalar_nan(value, result):
     assert is_scalar_nan(value) is result
-
-
-def test_init_arpack_v0():
-    v0s = []
-    for i in range(100):
-        v0s.append(_init_arpack_v0(1000, i))
-        if i > 0:
-            assert not any(np.equal(v0s[i], v0s[i-1]))
-
-    v0 = np.concatenate(v0s)
-    assert np.allclose(np.mean(v0), 0, atol=1e-2)
-    assert np.allclose(np.std(v0), 1/np.sqrt(3), atol=1e-3)
